api/pyproject.toml
[project]
name = "mealio-server"
version = "1.0.0"
description = "FastAPI backend for Mealio meal tracking application"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.16",
    "pydantic-settings>=2.5.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "twilio>=9.3.0",
    "pillow>=11.0.0",
    "python-magic>=0.4.27",
    "sentry-sdk[fastapi]>=2.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "greenlet>=3.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",
    "faker>=33.0.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501", "B008", "RUF012"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
